bug 558284 - replace com.spotify with org.mandas.docker-client
* also update target locations to https and fix name of target e4.16

Reviewed-on: https://git.eclipse.org/r/162605
Tested-by: Linux Tools Bot <linuxtools-bot@eclipse.org>
Tested-by: Jeff Johnston <jjohnstn@redhat.com>

+++ containers/org.eclipse.linuxtools.docker.core/src/org/eclipse/linuxtools/docker/core/IDockerContainerExit.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
- * Copyright (c) 2014, 2018 Red Hat.
- * 
+ * Copyright (c) 2014, 2020 Red Hat.
+ *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
  * which is available at https://www.eclipse.org/legal/epl-2.0/
@@ -14,6 +14,6 @@ package org.eclipse.linuxtools.docker.core;
 
 public interface IDockerContainerExit {
 
-	Integer statusCode();
+	Long statusCode();
 
 }
