+++ containers/org.eclipse.linuxtools.docker.core/src/org/eclipse/linuxtools/internal/docker/core/DockerContainerState.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
- * Copyright (c) 2014, 2018 Red Hat.
- * 
+ * Copyright (c) 2014, 2020 Red Hat.
+ *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
  * which is available at https://www.eclipse.org/legal/epl-2.0/
@@ -15,8 +15,7 @@ package org.eclipse.linuxtools.internal.docker.core;
 import java.util.Date;
 
 import org.eclipse.linuxtools.docker.core.IDockerContainerState;
-
-import com.spotify.docker.client.messages.ContainerState;
+import org.mandas.docker.client.messages.ContainerState;
 
 public class DockerContainerState implements IDockerContainerState {
 
@@ -24,7 +23,7 @@ public class DockerContainerState implements IDockerContainerState {
 	private final Boolean running;
 	private final Boolean paused;
 	private final Boolean restarting;
-	private final Integer exitCode;
+	private final Long exitCode;
 	private final Date startDate;
 	private final Date finishDate;
 
@@ -47,19 +46,19 @@ public class DockerContainerState implements IDockerContainerState {
 	public Boolean restarting() {
 		return restarting;
 	}
-	
+
 	@Override
 	public Boolean paused() {
 		return paused;
 	}
-	
+
 	@Override
 	public Integer pid() {
 		return pid;
 	}
 
 	@Override
-	public Integer exitCode() {
+	public Long exitCode() {
 		return exitCode;
 	}
