+++ containers/org.eclipse.linuxtools.docker.ui/src/org/eclipse/linuxtools/internal/docker/ui/launch/ContainerCommandProcess.java
@@ -1,6 +1,6 @@
 /*******************************************************************************
- * Copyright (c) 2017, 2019 Red Hat Inc. and others.
- * 
+ * Copyright (c) 2017, 2020 Red Hat Inc. and others.
+ *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
  * which is available at https://www.eclipse.org/legal/epl-2.0/
@@ -221,7 +221,7 @@ public class ContainerCommandProcess extends Process {
 							LaunchMessages.getFormattedString(
 									"ContainerNotFinished.msg", containerId)); //$NON-NLS-1$
 				}
-				return state.exitCode();
+				return state.exitCode().intValue();
 			}
 		}
 		if (containerRemoved) {
@@ -276,7 +276,7 @@ public class ContainerCommandProcess extends Process {
 				containerRemoved = true;
 				connection.removeContainer(containerId);
 			}
-			return exit.statusCode();
+			return exit.statusCode().intValue();
 		} catch (DockerException e) {
 			return -1;
 		}
